[Edit] AlarmStoreView - 유저이름 외 cell 눌러서 post 이동 #61
