+++ FinalHrmiProjects/View/Mypage/AlarmStoreView.swift
@@ -33,7 +33,7 @@ struct Alarm: Identifiable, Hashable {
         .init(likedUserName: "phang", postImageName: "foodEx2", likedTime: randomDate()),
         .init(likedUserName: "mangJae", postImageName: "foodEx4", likedTime: randomDate()),
         .init(likedUserName: "phang", postImageName: "foodEx4", likedTime: randomDate()),
-        .init(likedUserName: "내가아이디가좀길어", postImageName: "foodEx4", likedTime: randomDate()),
+        .init(likedUserName: "내가아이디가좀길어ㅋ", postImageName: "foodEx4", likedTime: randomDate()),
     ]
 }
 
@@ -45,7 +45,12 @@ struct AlarmStoreView: View {
             // TODO: CustomScrollView 로 수정 예정
             ScrollView {
                 ForEach(0..<Alarm.alarmList.count, id: \.self) { index in
-                    AlarmStoreListCell(alarm: Alarm.alarmList[index])
+                    NavigationLink {
+                        // TODO: 해당 게시글로 이동
+                    } label: {
+                        AlarmStoreListCell(alarm: Alarm.alarmList[index])
+                    }
+                    
                     if index != Alarm.alarmList.count - 1 {
                         CustomDivider()
                     }
@@ -71,52 +76,6 @@ struct AlarmStoreView: View {
                     .foregroundStyle(.mainBlack)
             }
         }
-    }
-}
-
-struct AlarmStoreListCell: View {
-    let alarm: Alarm
-    
-    var body: some View {
-        HStack(alignment: .center) {
-            VStack(alignment: .leading, spacing: 2) {
-                Group {
-                    Text(alarm.likedUserName)
-                        .font(.medium14)
-                    +
-                    Text(" 님이 게시물에 하트를 남겼어요.")
-                        .font(.regular14)
-                }
-                .foregroundStyle(.mainBlack)
-                .overlay(alignment: .topLeading) {
-                    NavigationLink {
-                        // TODO: 해당 유저 프로필로 이동
-                    } label: {
-                        Text(alarm.likedUserName)
-                            .font(.medium14)
-                            .foregroundStyle(.mainBlack)
-                    }
-                    .buttonStyle(EmptyActionStyle())
-                }
-                Text(Formatter.formattedDateBeforeStyle(pastDate: alarm.likedTime))
-                    .font(.regular12)
-                    .foregroundStyle(.gray01)
-            }
-            Spacer()
-            NavigationLink {
-                // TODO: 해당 게시글로 이동
-            } label: {
-                Image(alarm.postImageName)
-                    .resizable()
-                    .aspectRatio(1.0, contentMode: .fill)
-                    .frame(maxWidth: 60, maxHeight: 60)
-                    .clipShape(.rect(cornerRadius: 5))
-            }
-            .buttonStyle(EmptyActionStyle())
-        }
-        .padding(.horizontal, 10)
-        .frame(maxWidth: .infinity, maxHeight: 80, alignment: .leading)
-        .padding(.horizontal, 20)
     }
 }
